files/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pyotp==2.9.0
python-multipart==0.0.9
httpx==0.27.0
requests==2.32.3
python-dotenv==1.0.1
pydantic[email]==2.7.1
pydantic-settings==2.2.1
qrcode[pil]==7.4.2
cryptography==42.0.7
